refactor: consolidate providers into single index.ts entry point
- Add root index.ts that loads all 11 providers
- Move each provider into its own subdirectory under providers/
- Update package.json to point to ./index.ts
- Update test imports to reference new subdirectory paths
- Fix import paths for relocated files
