+++ package.json
@@ -35,6 +35,7 @@
 		"node": ">=20.0.0"
 	},
 	"files": [
+		"index.ts",
 		"providers/**/*.ts",
 		"lib/**/*.ts",
 		"usage/**/*.ts",
@@ -69,17 +70,7 @@
 	},
 	"pi": {
 		"extensions": [
-			"./providers/kilo.ts",
-			"./providers/zen.ts",
-			"./providers/go.ts",
-			"./providers/openrouter.ts",
-			"./providers/nvidia.ts",
-			"./providers/cline.ts",
-			"./providers/fireworks.ts",
-			"./providers/mistral.ts",
-			"./providers/ollama.ts",
-			"./providers/qwen.ts",
-			"./providers/modal.ts"
+			"./index.ts"
 		]
 	}
 }
