+++ providers/cline/cline.ts
@@ -14,9 +14,9 @@
 
 import type { OAuthCredentials } from "@mariozechner/pi-ai";
 import type { ExtensionAPI } from "@mariozechner/pi-coding-agent";
-import { BASE_URL_CLINE, PROVIDER_CLINE } from "../constants.ts";
-import { incrementRequestCount } from "../usage/metrics.ts";
-import { logWarning } from "../lib/util.ts";
+import { BASE_URL_CLINE, PROVIDER_CLINE } from "../../constants.ts";
+import { incrementRequestCount } from "../../usage/metrics.ts";
+import { logWarning } from "../../lib/util.ts";
 import { loginCline, refreshClineToken } from "./cline-auth.ts";
 import { fetchClineModels } from "./cline-models.ts";
