+++ providers/kilo/kilo-auth.ts
@@ -9,9 +9,9 @@ import type {
 import {
 	KILO_POLL_INTERVAL_MS,
 	KILO_TOKEN_EXPIRATION_MS,
-} from "../constants.ts";
-import { createLogger } from "../lib/logger.ts";
-import { openBrowser } from "../lib/open-browser.ts";
+} from "../../constants.ts";
+import { createLogger } from "../../lib/logger.ts";
+import { openBrowser } from "../../lib/open-browser.ts";
 
 const _logger = createLogger("kilo-auth");
