+++ providers/kilo/kilo-models.ts
@@ -2,8 +2,8 @@
  * Kilo model fetching and mapping (OpenRouter-compatible format).
  */
 
-import { applyHidden } from "../config.ts";
-import { fetchOpenRouterCompatibleModels } from "./model-fetcher.ts";
+import { applyHidden } from "../../config.ts";
+import { fetchOpenRouterCompatibleModels } from "../model-fetcher.ts";
 
 const KILO_API_BASE = process.env.KILO_API_URL || "https://api.kilo.ai";
 export const KILO_GATEWAY_BASE = `${KILO_API_BASE}/api/gateway`;
